Html Code.
To get started, we will first need to create a basic HTML file. In this file, we will include the main structure for our calculator.


We will start by creating a div element and giving it the class of container. This div will act as a container for our calculator.
Next, we will add buttons and an input field to the calculator. We will use the button element to create the buttons, and the input element to create the input field.


After creating the files just paste the following codes into your file. Make sure to save your HTML document with a .html extension, so that it can be properly viewed in a web browser.
This is the basic structure of our calculator using HTML, and now we can move on to styling it using CSS.